move storage into storage pkg
